pyproject.toml
[project]
name = "python-e-commerce"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "black>=26.3.1",
    "isort>=8.0.1",
    "langchain>=1.2.13",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.12",
    "python-dotenv>=1.2.2",
]
